+++ packages/date-picker/docs/filter-button-foundations/FilterBaseButton.stories.tsx
@@ -1,6 +1,7 @@
 import React from "react"
-import { ComponentMeta, ComponentStory } from "@storybook/react"
+import { ComponentMeta, ComponentStory, Story } from "@storybook/react"
 import { withDesign } from "storybook-addon-designs"
+import { StickerSheet } from "../../../../storybook/components/StickerSheet"
 import {
   CATEGORIES,
   SUB_CATEGORIES,
@@ -28,4 +29,31 @@ export const FilterBaseButtonStory: ComponentStory<
 FilterBaseButtonStory.storyName = "Filter Base Button"
 FilterBaseButtonStory.args = {
   children: "Label",
+}
+
+const StickerSheetTemplate: Story = () => (
+  <StickerSheet heading="Filter Base Button">
+    <StickerSheet.Header headings={["Base", "Hover", "Active", "Focus"]} />
+    <StickerSheet.Body>
+      <StickerSheet.Row>
+        <FilterBaseButton>Label</FilterBaseButton>
+        <FilterBaseButton classNameOverride="story__filter-button--hover">
+          Label
+        </FilterBaseButton>
+        <FilterBaseButton classNameOverride="story__filter-button--active">
+          Label
+        </FilterBaseButton>
+        <FilterBaseButton classNameOverride="story__filter-button--focus">
+          Label
+        </FilterBaseButton>
+      </StickerSheet.Row>
+    </StickerSheet.Body>
+  </StickerSheet>
+)
+
+export const StickerSheetDefault = StickerSheetTemplate.bind({})
+StickerSheetDefault.storyName = "Sticker Sheet (Default)"
+StickerSheetDefault.parameters = {
+  chromatic: { disable: false },
+  controls: { disable: true },
 }