+++ packages/select/src/Select/components/index.ts
@@ -1,1 +1,3 @@
-export type { SingleTriggerButtonProps } from "./Trigger"
+export type { TriggerButtonProps } from "./TriggerButton"
+export type { OptionProps } from "./Option"
+export type { SingleListBoxProps } from "./ListBox"